Moving new material views to base class
