+++ app/src/main/res/layout/download_bottom_sheet.xml
@@ -76,7 +76,7 @@
             android:layout_gravity="center"
             android:visibility="gone" />
 
-        <eu.kanade.tachiyomi.ui.library.MaterialFastScroll
+        <eu.kanade.tachiyomi.ui.base.MaterialFastScroll
             android:id="@+id/fast_scroller"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
